ai-service/requirements.txt
flask
kafka-python
scikit-learn
pandas
